+++ lib/Control/Monad/Dep/Advice.hs
@@ -1,7 +1,0 @@
-module Control.Monad.Dep.Advice {-# DEPRECATED "Renamed to Dep.Advice" #-}
-  (
-    module Dep.Advice
-  )
-where
-
-import Dep.Advice