+++ lib/Control/Monad/Dep/Advice/Basic.hs
@@ -1,7 +1,0 @@
-module Control.Monad.Dep.Advice.Basic {-# DEPRECATED "Renamed to Dep.Advice.Basic" #-}
-  ( 
-    module Dep.Advice.Basic
-  )
-where
-
-import Dep.Advice.Basic