+++ lib/Control/Monad/Dep/SimpleAdvice.hs
@@ -1,8 +1,0 @@
-module Control.Monad.Dep.SimpleAdvice {-# DEPRECATED "Renamed to Dep.SimpleAdvice" #-}
-  (
-    module Dep.SimpleAdvice
-  )
-where
-
-import Dep.SimpleAdvice
-