+++ lib/Control/Monad/Dep/SimpleAdvice/Basic.hs
@@ -1,7 +1,0 @@
-module Control.Monad.Dep.SimpleAdvice.Basic {-# DEPRECATED "Renamed to Dep.SimpleAdvice.Basic" #-}
-  (
-    module Dep.SimpleAdvice.Basic 
-  )
-where
-
-import Dep.SimpleAdvice.Basic